+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ### [TURISTEA-WEB]
 
+### [0.0.2] - 15-08-2025 
+-Add service login
 
 ### [0.0.1] - 04-08-2025 
 ### Feature
 - Comming template admin 
+- Delete archives not necessary